APIs/azure.com/monitor-logProfiles_API/2016-03-01/python-aiohttp-client/tests/test_log_profiles_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.error_response import ErrorResponse
from openapi_server.models.log_profile_collection import LogProfileCollection
from openapi_server.models.log_profile_resource import LogProfileResource


pytestmark = pytest.mark.asyncio

async def test_log_profiles_create_or_update(client):
    """Test case for log_profiles_create_or_update

    
    """
    parameters = {"properties":{"serviceBusRuleId":"serviceBusRuleId","storageAccountId":"storageAccountId","locations":["locations","locations"],"categories":["categories","categories"],"retentionPolicy":{"days":0,"enabled":True}}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/subscriptions/{subscription_id}/providers/microsoft.insights/logprofiles/{log_profile_name}'.format(log_profile_name='log_profile_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_log_profiles_delete(client):
    """Test case for log_profiles_delete

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/subscriptions/{subscription_id}/providers/microsoft.insights/logprofiles/{log_profile_name}'.format(log_profile_name='log_profile_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_log_profiles_get(client):
    """Test case for log_profiles_get

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/providers/microsoft.insights/logprofiles/{log_profile_name}'.format(log_profile_name='log_profile_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_log_profiles_list(client):
    """Test case for log_profiles_list

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/providers/microsoft.insights/logprofiles'.format(subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

